Bazel: offload BUILD files
This commit introduces some basic BUILD files to offload the main Bazel branch. It runs Bazel via Bazelisk on CI and currently executes only two jobs: Buldifier to lint BUILD and WORKSPACE files and `bazel build`.

It should not be used yet so there is no documentation for it yet. I will keep offloading things in a non-breaking fashion in other commits.

adeira-source-id: 513b95d56fd4f4dc186e13fdd92cb7bb0473a63c